situation test6

+++ src/main/java/com/webank/webase/front/situation/entity/Situation.java
@@ -21,6 +21,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigInteger;
 
 /**
  * entity of Node Monitor in db
@@ -33,7 +34,7 @@ public class Situation {
     @GeneratedValue(strategy= GenerationType.AUTO)
     private Long id;
 
-    private int txPool;
+    private BigInteger txPool;
     private int sealer;
     private int consensusEngineBlock;
     private int consensusEngineCommonView;
